Support a SQL Like() function
RegEx's had take unlimited time, so supporting them against the DB can be hazardous.
LIKE(String,String) does a SQL LIKE supporting '%' and '_' in a case insensitive fashion in order
  to allow EndsWith functionality without doing unnatural and expensive things using REVERSE.
  Backslash \ is used as the escape character
In javascript and java, it is converted to a regular expression.
Oracle doesn't like invalid escape characters and Sqlite3 is case insensitive by default and can't
  be changed except globally, but works most other places.