@W-14808875 - allow ISNULL to work with booleans when used in clientside JavaScript
This is needed in Screen Flows for checkbox inputs which can be hidden. If a field is hidden by rules then it's value should be null. This is true for all fields, but unfortunately checkboxes are immediately registed as booleans and therefore due to the formula engine restricting null checks on booleans we have a number of bugs that come up.

This fix allows booleans to be null checked when the allowNullableBooleansInJs flag is set to true.